requirements.txt
docker
pyyaml
requests
aliyun-python-sdk-core
aliyun-python-sdk-cr
